move helper to describe block
